requirements.txt
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
transformers
sentence-transformers
numpy
langchain
langchain-core
langchain-community
langchain_huggingface
langchain-chroma
langchain-groq
langchain-openai
groq
openai
streamlit
beautifulsoup4
requests
chromadb
python-dotenv
